Added request logging filters. Closes #30. Refactored all package names to clean up the folder structure.
