+++ app/core/localization/Constants.scala
@@ -1,0 +1,14 @@
+package core.localization
+
+/**
+  * Common message constants used across the application
+  */
+object Constants {
+
+  /**
+    * Used for the correlation id's to match a full request flow.
+    */
+  val correlationIdHeader = "CorrelationId"
+
+  val applicationName = "Openwalrus"
+}
