--- app/filters/ApplicationFilters.scala
@@ -1,13 +1,0 @@
-package filters
-
-import javax.inject.Inject
-
-import play.api.http.HttpFilters
-import play.filters.gzip.GzipFilter
-
-/**
-  * Common filters to apply across application.
-  */
-class ApplicationFilters @Inject() (gzipFilter: GzipFilter) extends HttpFilters {
-  def filters = Seq(gzipFilter)
-}
