+++ conf/application.conf
@@ -15,7 +15,7 @@ play.modules.enabled += "play.modules.reactivemongo.ReactiveMongoModule"
 
 # Special
 # ~~~~~
-play.http.filters = "filters.ApplicationFilters"
+play.http.filters = "core.filters.ApplicationFilters"
 # application.context="/openwalrus"
 
 # Data
