+++ test/business/ManagerTestBase.scala
@@ -1,13 +1,10 @@
 package business
 
+import core.BaseTestSpec
+import data.daos.UserDao
 import play.api.inject.bind
-import walrath.technology.openwalrus.daos.{FileDao, UserDao}
-import org.scalamock.scalatest.MockFactory
-import org.scalatest.{BeforeAndAfter, Matchers, WordSpec }
-import play.api.Mode
+import data.daos.FileDao
 import play.api.inject.guice.GuiceInjectorBuilder
-import org.scalatest.OneInstancePerTest
-import walrath.technology.openwalrus.testing.BaseTestSpec
 
 trait ManagerTestBase extends BaseTestSpec {
   val userDaoMock = mock[UserDao]
