+++ test/core/BaseTestSpec.scala
@@ -1,0 +1,10 @@
+package core
+
+import org.scalatest.WordSpec
+import org.scalatest.BeforeAndAfter
+import org.scalatest.Matchers
+import org.scalamock.scalatest.MockFactory
+import org.scalatest.OneInstancePerTest
+
+
+trait BaseTestSpec extends WordSpec with BeforeAndAfter with Matchers with MockFactory with OneInstancePerTest
